ember-tooltip event handling working for ember-tooltip component
