+++ addon/templates/components/ember-tooltip-simple.hbs
@@ -1,7 +1,7 @@
 {{#if hasBlock}}
   {{#ember-wormhole
     to=wormholeId
-    renderInPlace=_renderWormholeInPlace
+    renderInPlace=_tooltipElementNotRendered
   }}
     <div>
       {{yield}}
