+++ tests/dummy/app/templates/index.hbs
@@ -1,19 +1,9 @@
 <br>
-<span id="new-test">
-  Hey
-  <br>
-  {{#ember-tooltip side='right'}}
-    Test
-  {{/ember-tooltip}}
-  <br>
-</span>
-
-<br>
 <br>
 
-<span id="test2">
+<span>
   Sup
-  {{#ember-tooltip-simple}}
+  {{#ember-tooltip}}
     Ok
-  {{/ember-tooltip-simple}}
+  {{/ember-tooltip}}
 </span>
